move metadb into internal
