+++ internal/metadb/gauges.go
