+++ tools/amplitude/build.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/GeoNet/delta/metadb"
+	"github.com/GeoNet/delta/internal/metadb"
 	"github.com/GeoNet/delta/resp"
 )
