+++ tools/chart/networks.go
@@ -4,8 +4,8 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/GeoNet/delta/internal/metadb"
 	"github.com/GeoNet/delta/meta"
-	"github.com/GeoNet/delta/metadb"
 )
 
 func (cp ConfigPage) Networks(base string) ([]Page, error) {
